requirements.txt
streamlit
azure-cognitiveservices-vision-customvision
msrest==0.7.1
Pillow
